feat: separate threshold maker from detectors
Filename changed, update related import paths.
